Initial implementation of render-to-texture
