+++ src/surface.rs
@@ -1,0 +1,9 @@
+pub struct Surface {
+    surface: golem::Surface
+}
+
+impl Surface {
+    pub fn new(gfx: &Graphics, attachment: Image) -> Surface {
+
+    }
+}
